pyproject.toml
[project]
name = "flowcast"
version = "0.1.0"
description = "Convert human instructions to GIFs"
requires-python = ">=3.11"
dependencies = [
    "pyautogui>=0.9.54",
    "pillow>=10.3.0",
    "python-dotenv>=1.0.1",
    "pyobjc-framework-Quartz>=10.3.1",
    "pyobjc-core>=10.3.1",
    "groq>=0.13.0",
    "easyocr>=1.7.1",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.8.0",
    "scipy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
